% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_conferences}
\alias{dictionary_conferences}
\title{Data Dictionary: Conferences}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\usage{
dictionary_conferences
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_conferences]{load_conferences()}}
}
\keyword{datasets}
