% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_games}
\alias{dictionary_games}
\title{Data Dictionary: Games}
\format{
An object of class \code{data.frame} with 47 rows and 3 columns.
}
\usage{
dictionary_games
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_games]{load_games()}}
}
\keyword{datasets}
