% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_lines}
\alias{dictionary_lines}
\title{Data Dictionary: Lines}
\format{
An object of class \code{data.frame} with 19 rows and 3 columns.
}
\usage{
dictionary_lines
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_lines]{load_lines()}}
}
\keyword{datasets}
