% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_media}
\alias{dictionary_media}
\title{Data Dictionary: Media}
\format{
An object of class \code{data.frame} with 18 rows and 3 columns.
}
\usage{
dictionary_media
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_media]{load_media()}}
}
\keyword{datasets}
