% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_player_box_scores}
\alias{dictionary_player_box_scores}
\title{Data Dictionary: Player Box Scores}
\format{
An object of class \code{data.frame} with 60 rows and 3 columns.
}
\usage{
dictionary_player_box_scores
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_player_box_scores]{load_player_box_scores()}}
}
\keyword{datasets}
