% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_plays}
\alias{dictionary_plays}
\title{Data Dictionary: Plays}
\format{
An object of class \code{data.frame} with 86 rows and 3 columns.
}
\usage{
dictionary_plays
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_plays]{load_plays()}}
}
\keyword{datasets}
