% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_rankings}
\alias{dictionary_rankings}
\title{Data Dictionary: Rankings}
\format{
An object of class \code{data.frame} with 11 rows and 3 columns.
}
\usage{
dictionary_rankings
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_rankings]{load_rankings()}}
}
\keyword{datasets}
