% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_recruiting}
\alias{dictionary_recruiting}
\title{Data Dictionary: Recruiting}
\format{
An object of class \code{data.frame} with 22 rows and 3 columns.
}
\usage{
dictionary_recruiting
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_recruiting]{load_recruiting()}}
}
\keyword{datasets}
