% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_team_box_scores}
\alias{dictionary_team_box_scores}
\title{Data Dictionary: Team Box Scores}
\format{
An object of class \code{data.frame} with 108 rows and 3 columns.
}
\usage{
dictionary_team_box_scores
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_team_box_scores]{load_team_box_scores()}}
}
\keyword{datasets}
