% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_teams}
\alias{dictionary_teams}
\title{Data Dictionary: Teams}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\usage{
dictionary_teams
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_teams]{load_teams()}}
}
\keyword{datasets}
