% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictionary_venues}
\alias{dictionary_venues}
\title{Data Dictionary: Venues}
\format{
An object of class \code{data.frame} with 6 rows and 3 columns.
}
\usage{
dictionary_venues
}
\description{
A dataframe containing the data dictionary for \code{\link[=load_venues]{load_venues()}}
}
\keyword{datasets}
