% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_conferences.R
\name{load_conferences}
\alias{load_conferences}
\title{Load cbb conference information}
\usage{
load_conferences()
}
\value{
a dataframe of information about college basketball conferences.
}
\description{
Loads information about college basketball conferences
}
\examples{
\donttest{
try({ # avoid cran errors
load_conferences()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_conferences.html} for a web version of the data dictionary

\code{\link{dictionary_conferences}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
