% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lines.R
\name{load_lines}
\alias{load_lines}
\title{Load betting information for cbb games}
\usage{
load_lines(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of betting information for all games for the specified season(s).
}
\description{
Pull in vegas odds (overs, unders, spreads, moneylines, etc.) for college basketball games.
}
\examples{
\donttest{
try({ # avoid cran errors
load_lines()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_lines.html} for a web version of the data dictionary

\code{\link{dictionary_lines}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
