% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_media.R
\name{load_media}
\alias{load_media}
\title{Load media for cbb games}
\usage{
load_media(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of broadcast information for all games for the specified season(s).
}
\description{
Pull in broadcast information for all games in a given season
}
\examples{
\donttest{
try({ # avoid cran errors
load_media()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_media.html} for a web version of the data dictionary

\code{\link{dictionary_media}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
