% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rankings.R
\name{load_rankings}
\alias{load_rankings}
\title{Load ranking data}
\usage{
load_rankings()
}
\value{
a dataframe of information about college basketball rankings and polls
}
\description{
Loads information about college basketball rankings, polls, etc.

At present, only the AP Top 25 and Coaches Polls are available from this function. Ranking, first place votes, and over points are available for each week of the season dating back to 1949.
}
\examples{
\donttest{
try({ # avoid cran errors
load_rankings()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_rankings.html} for a web version of the data dictionary

\code{\link{dictionary_rankings}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
