% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recruiting.R
\name{load_recruiting}
\alias{load_recruiting}
\title{Load recruiting data}
\usage{
load_recruiting()
}
\value{
a dataframe of information about college basketball recruits
}
\description{
Loads information about college basketball recruiting
}
\examples{
\donttest{
try({ # avoid cran errors
load_recruiting()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_recruiting.html} for a web version of the data dictionary

\code{\link{dictionary_recruiting}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
