% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rosters.R
\name{load_rosters}
\alias{load_rosters}
\title{Load cbb team rosters}
\usage{
load_rosters(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of rosters for college basketball teams in the specified season(s).
}
\description{
Pull in rosters for college basketball teams for specified season(s)
}
\examples{
\donttest{
try({ # avoid cran errors
load_rosters()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_rosters.html} for a web version of the data dictionary

\code{\link{dictionary_rosters}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
