% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_team_stats.R
\name{load_team_stats}
\alias{load_team_stats}
\title{Load cbb teams stats}
\usage{
load_team_stats(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer or vector of integers of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of team statistics for college basketball teams.
}
\description{
Pull in statistics aggregated by team for for specified season(s)
}
\examples{
\donttest{
try({ # avoid cran errors
load_team_stats()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_team_stats.html} for a web version of the data dictionary

\code{\link{dictionary_team_stats}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
