% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_venues.R
\name{load_venues}
\alias{load_venues}
\title{Load cbb venue information}
\usage{
load_venues()
}
\value{
a dataframe of information about college basketball venues.
}
\description{
Loads information about college basketball venues.
}
\examples{
\donttest{
try({ # avoid cran errors
load_venues()
})
}

}
\seealso{
\url{https://john-b-edwards.github.io/cbbreadr/articles/dictionary_venues.html} for a web version of the data dictionary

\code{\link{dictionary_venues}} for the data dictionary as bundled within the package

Issues with this data should be filed here: \url{https://github.com/john-b-edwards/cbbd-data/issues}
}
