% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{most_recent_season}
\alias{most_recent_season}
\title{get the most recent CBB season}
\usage{
most_recent_season(date = Sys.Date())
}
\arguments{
\item{date}{the date to obtain the most recent season of. Defaults to the system date.}
}
\value{
the most recent CBB season as an integer
}
\description{
For a given date (defaults to the system date), returns the most recent or "current" college basketball season.
}
\examples{
\donttest{
most_recent_season()

most_recent_season("2022-09-01") # up to september, will be recorded as 2022

most_recent_season("2022-10-01") # then will swap over to 2023 when we hit october
}

}
