% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_conferences.R
\name{load_conferences}
\alias{load_conferences}
\title{Load cbb conference information}
\usage{
load_conferences()
}
\value{
a dataframe of information about college basketball conferences.
}
\description{
Loads information about college basketball conferences
}
\examples{
\donttest{
load_conferences()
}
}
