% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_games.R
\name{load_games}
\alias{load_games}
\title{Load college basketball games for a given season}
\usage{
load_games(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of all games in the specified season(s).
}
\description{
Pull in all college basketball games tracked by collegebasketballdata.com for a given season
}
\examples{
\donttest{
load_games()

load_games(2024)

load_games(TRUE)
}

}
