% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_media.R
\name{load_media}
\alias{load_media}
\title{Load media for cbb games}
\usage{
load_media(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of broadcast information for all games for the specified season(s).
}
\description{
Pull in broadcast information for all games in a given season
}
\examples{
\donttest{
load_media()

load_media(2024)

load_media(TRUE)
}

}
