% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_player_box_scores.R
\name{load_player_box_scores}
\alias{load_player_box_scores}
\title{Load cbb player box score stats}
\usage{
load_player_box_scores(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of player box scores.
}
\description{
Pull in statistics aggregated by player and game for a given season.
}
\examples{
\donttest{
load_player_box_scores()

load_player_box_scores(2024)

load_player_box_scores(TRUE)
}

}
