% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_player_stats.R
\name{load_player_stats}
\alias{load_player_stats}
\title{Load cbb player stats}
\usage{
load_player_stats(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of player statistics for college basketball players
}
\description{
Pull in statistics aggregated by player for a given season.
}
\examples{
\donttest{
load_player_stats()

load_player_stats(2024)

load_player_stats(TRUE)
}

}
