% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_plays.R
\name{load_plays}
\alias{load_plays}
\title{Load cbb play by play data}
\usage{
load_plays(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of play-by-play data from all D1 college basketball games.
}
\description{
Grabs entire seasons of play-by-play data from the collegebasketballdata.com API.
}
\examples{
\donttest{
load_plays(2024)
}

}
