% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rankings.R
\name{load_rankings}
\alias{load_rankings}
\title{Load ranking data}
\usage{
load_rankings()
}
\value{
a dataframe of information about college basketball rankings and polls
}
\description{
Loads information about college basketball rankings, polls, etc.
}
\examples{
\donttest{
load_rankings()
}
}
