% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_recruiting.R
\name{load_recruiting}
\alias{load_recruiting}
\title{Load recruiting data}
\usage{
load_recruiting()
}
\value{
a dataframe of information about college basketball recruits
}
\description{
Loads information about college basketball recruiting
}
\examples{
\donttest{
load_recruiting()
}
}
