% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rosters.R
\name{load_rosters}
\alias{load_rosters}
\title{Load cbb team rosters}
\usage{
load_rosters(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of rosters for college basketball teams in the specified season(s).
}
\description{
Pull in rosters for college basketball teams for a given season
}
\examples{
\donttest{
load_rosters()

load_rosters(2024)

load_rosters(TRUE)
}

}
