% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_team_box_scores.R
\name{load_team_box_scores}
\alias{load_team_box_scores}
\title{Load cbb teams box score stats}
\usage{
load_team_box_scores(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of team box scores.
}
\description{
Pull in statistics aggregated by team and game for a given season.
}
\examples{
\donttest{
load_team_box_scores()

load_team_box_scores(2024)

load_team_box_scores(TRUE)
}

}
