% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_team_stats.R
\name{load_team_stats}
\alias{load_team_stats}
\title{Load cbb teams stats}
\usage{
load_team_stats(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of team statistics for college basketball teams.
}
\description{
Pull in statistics aggregated by team for a given season.
}
\examples{
\donttest{
load_team_stats()

load_team_stats(2024)

load_team_stats(TRUE)
}

}
