% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_teams.R
\name{load_teams}
\alias{load_teams}
\title{Load cbb teams}
\usage{
load_teams(seasons = most_recent_season())
}
\arguments{
\item{seasons}{an integer of seasons to fetch data for. Defaults to the most recent season. Pass in \code{TRUE} to fetch all seasons.}
}
\value{
a dataframe of all college basketball teams tracked by collegebasketballdata.com for a given season(s).
}
\description{
Pull in information on all teams that played in a given season.
}
\examples{
\donttest{
load_team()

load_team(2024)

load_teams(TRUE)
}

}
