% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_venues.R
\name{load_venues}
\alias{load_venues}
\title{Load cbb venue information}
\usage{
load_venues()
}
\value{
a dataframe of information about college basketball venues.
}
\description{
Loads information about college basketball venues.
}
\examples{
\donttest{
load_venues()
}
}
